Fix invoicing numbers serializing as strings, not numbers
Pydantic's default JSON encoding sends Decimal fields as strings, which
crashed every frontend screen calling .toFixed() on them. Response schemas
now use float instead — DB columns stay Numeric for storage precision.

+++ app/schemas.py
@@ -3,7 +3,6 @@ import uuid
 from datetime import date, datetime, time
 from typing import Optional
 from pydantic import BaseModel, ConfigDict
-from decimal import Decimal
 from .models import (AvailabilityStatus, ChatChannelType, CheckInStatus,
                      CoverageType, InvoiceStatus, LicenceStatus, OnboardingStatus,
                      OperatorRole, ShiftStatus, SiteStatus, SiteType)
@@ -60,7 +59,7 @@ class OperatorOut(BaseModel):
     security_licence_expiry: Optional[date] = None
     licence_status: Optional[LicenceStatus] = None
     # Contractor pay rate — same visibility rule as the licence fields above.
-    pay_rate: Optional[Decimal] = None
+    pay_rate: Optional[float] = None
 
 
 class OperatorCreate(BaseModel):
@@ -90,11 +89,11 @@ class SiteOut(BaseModel):
     color: str
     active: bool
     # Admin/Director only — never sent on the operator-facing site card.
-    bill_rate: Optional[Decimal] = None
+    bill_rate: Optional[float] = None
 
 
 class RatePatch(BaseModel):
-    rate: Optional[Decimal] = None
+    rate: Optional[float] = None
 
 
 # ── Assignment ────────────────────────────────────────────────────────────────
@@ -219,7 +218,7 @@ class OperatorWithRoles(BaseModel):
     roles: list[OperatorRoleOut] = []
     site_accesses: list[SiteAccessOut] = []
     # This listing is admin-only already, so the rate is safe to include.
-    pay_rate: Optional[Decimal] = None
+    pay_rate: Optional[float] = None
 
 
 class AssignRoleRequest(BaseModel):
@@ -587,9 +586,9 @@ class InvoiceLineItemOut(BaseModel):
     shift_name: str
     start_time: Optional[time] = None
     end_time: Optional[time] = None
-    hours: Decimal
-    rate: Decimal
-    amount: Decimal
+    hours: float
+    rate: float
+    amount: float
 
 
 class InvoiceOut(BaseModel):
@@ -605,9 +604,9 @@ class InvoiceOut(BaseModel):
     approved_by: Optional[uuid.UUID] = None
     paid_at: Optional[datetime] = None
     marked_paid_by: Optional[uuid.UUID] = None
-    total_hours: Decimal
-    total_amount: Decimal
-    gst_amount: Optional[Decimal] = None
+    total_hours: float
+    total_amount: float
+    gst_amount: Optional[float] = None
     created_at: datetime
 
 
@@ -619,8 +618,8 @@ class InvoicePreviewOut(BaseModel):
     """Not persisted — a projection of what submitting would create."""
     period_month: int
     period_year: int
-    total_hours: Decimal
-    total_amount: Decimal
+    total_hours: float
+    total_amount: float
     line_items: list[InvoiceLineItemOut] = []
 
 
@@ -630,10 +629,10 @@ class SiteFinancialOut(BaseModel):
     site_slug: str
     period_month: int
     period_year: int
-    billable_hours: Decimal
-    bill_amount: Decimal
-    pay_amount: Decimal
-    margin: Decimal
+    billable_hours: float
+    bill_amount: float
+    pay_amount: float
+    margin: float
 
 
 class CoverageVarianceOut(BaseModel):
@@ -642,7 +641,7 @@ class CoverageVarianceOut(BaseModel):
     site_slug: str
     period_month: int
     period_year: int
-    scheduled_hours: Decimal
-    actual_hours: Decimal
-    variance_hours: Decimal
-    variance_pct: Optional[Decimal] = None
+    scheduled_hours: float
+    actual_hours: float
+    variance_hours: float
+    variance_pct: Optional[float] = None
